### 2024-09-28 42f7864
update CRAN version 1.1.0
