+++ DESCRIPTION
@@ -5,7 +5,7 @@ Version: 1.1.0
 Authors@R: c(
     person("Biying", "Zhou", email = "zby.zhou@mail.utoronto.ca", role = "cre"),
     person("Feng", "Ji", email = "f.ji@utoronto.ca", role = "aut"))
-Description: Integrate Item Response Theory (IRT) and Federated Learning to estimate traditional IRT models, including the 2-Parameter Logistic (2PL) and the Graded Response Models, with enhanced privacy. It allows for the estimation in a distributed manner without compromising accuracy. A user-friendly 'shiny' application is included. For more details, see Biying Zhou, Feng Ji (2024) "FedIRT: An R package and 'shiny' app for estimating federated item response theory models" <https://github.com/Feng-Ji-Lab/FedIRT/blob/main/paper/paper.pdf>.
+Description: Integrate Item Response Theory (IRT) and Federated Learning to estimate traditional IRT models, including the 2-Parameter Logistic (2PL) and the Graded Response Models, with enhanced privacy. It allows for the estimation in a distributed manner without compromising accuracy. A user-friendly 'shiny' application is included. 
 License: MIT + file LICENSE
 Depends:
     R (>= 3.1.0),
