### 2025-02-09 #7255 @salbalkus
@samhforbes yes, I'd be happy to review the package!
