### 2025-02-21 #7255 @asmitamitra
No, I don't have any previous experience or work in the Psychometrics/item response theory field.
