### 2025-03-03 #7255 @samhforbes
OK I think in the absence of other responses I will keep an eye on the psychometrics side of this as well. I'll look to begin the review so we aren't delayed, but I'll add another reviewer if someone with additional psychometrics understanding beyond me also agrees.
