### 2025-03-03 #7869 @editorialbot
[REVIEW]: FedIRT: An R package and Shiny app for estimating federated item response theory models
**Submitting author:** @trtfj (<a href="http://orcid.org/0000-0002-2051-5453">Feng Ji</a>)
**Repository:** https://github.com/Feng-Ji-Lab/FedIRT
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
