### 2025-03-04 Feng-Ji-Lab/FedIRT#1 @salbalkus
JOSS review: License
Raised as part of JOSS review openjournals/joss-reviews#7869.

Though you've included an MIT license in LICENSE.md, it appears that the LICENSE plaintext file does not contain a valid license. I think in order for the license to be detected by JOSS, it should be in the LICENSE plaintext file instead.
