### 2025-03-04 Feng-Ji-Lab/FedIRT#2 @salbalkus
JOSS review: Data
Raised as part of JOSS review openjournals/joss-reviews#7869.

I just want to confirm that the data you've included in this package complies with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies). Could you provide some information on where this data came from? Is it artificial, sampled from real students, or from other sources?
