### 2025-03-05 Feng-Ji-Lab/FedIRT#3 @salbalkus
JOSS review: Missing documentation
Raised as part of JOSS review https://github.com/openjournals/joss-reviews/issues/7869.

Most of the "Documentation" requirements appear to be missing. See the following for specifics:

**Statement of Need**: I couldn't find one. At minimum, top of README.md in the repo should tell visitors what the package is used for.

**Installation Instructions**: Can this package be installed via CRAN? If so, need to update the installation instructions to include the CRAN method.

**Example Usage**: A strong explanation of how to use the Shiny App part should be placed first, as to my understanding, this is the actual highlight of the package. Also, it might be worth setting up a github.io site to host documentation (i.e. via [pkgdown](https://pkgdown.r-lib.org/)) as this makes it easier for people to find and read your examples.

**Automated tests**: Are you able to run the tests automatically? For me, `test_check("FedIRT")` seems to only run the single example test in the `testthat` folder. You might need to move the tests you've written into that folder so that they can be automatically run.

**Community Guidelines**: Missing clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support, as required by JOSS.
