### 2025-03-05 Feng-Ji-Lab/FedIRT#4 @salbalkus
JOSS review: Paper
Raised as part of JOSS review https://github.com/openjournals/joss-reviews/issues/7869.

The paper does an overall good job of describing the package's functionality. I have some recommendations to improve certain elements and trim it down to the JOSS guidelines of ~1000 words, which I've broken up by part:

**Summary**: As someone not in psychometrics, but familiar with federated learning, I found that your description "beats around the bush" too much for a non-specialist audience. At a minimum, it needs at least one sentence explaining why we need this package. Why do we need to do federated learning for item response theory? I would assume it is needed so that social scientists can analyze educational testing data from school districts, as this data would otherwise be unavailable to use due to privacy concerns. A more focused, jargon-free summary would be much appreciated.

**Statement of Need**: I recommend some additions:
1. Need to explain what IRT is used for in non-specialist terms (non-specialists will not know what "measure latent constructs through multivariate human responses" means). Give some practical examples of how it is used.
2. Need to state clearly who this is for in the Statement of Need. Education researchers? social scientists? The school districts themselves? Are the researchers meant to convince school districts to let them use this software to analyze their data?
3. A bonus that might be worth adding: federated learning is a huge theoretical field, and I'm not an advanced expert, but as far as I know, this is the first open-source R package that actually *performs* the "distributed" part and can be used in practice. In the past, I've actually been quite skeptical of federated learning research precisely because few people ever provide software that does the distributed part, but this package bucks the trend in an excellent way. Might be worth mentioning if you want to strengthen the paper.

**Quality of Writing**: Good, but it appears that the word count of the paper is quite long. I think you could remove the "Method" section, or shorten it to one or two sentences saying that you use federated stochastic gradient descent to fit the model described in your paper (and move this to Statement of Need). Also, you should remove the "Availability", "Example of the integrated function", "Example of the personscore function", and "Example of the personfit function" from the paper, as these are documentation. I think you should keep the "Example of the Shiny App" though, as this is important context for readers to understand that this package actually performs the model distribution using Shiny, which is a huge step for the field.
