### 2025-03-05 #7869 @salbalkus
Hi @trtfj and @BiyingZhou,

Thank you for this nice addition to the R ecosystem. I have yet to check the functionality of the software, which I'll do a little later, so thus far I've gone through everything besides that. I've found four major areas that require addressing, which I've included as issues on the GitHub repo:

- [ ] [JOSS review: License Feng-Ji-Lab/FedIRT#1](https://github.com/Feng-Ji-Lab/FedIRT/issues/1)
- [ ] [JOSS review: Data Feng-Ji-Lab/FedIRT#2](https://github.com/Feng-Ji-Lab/FedIRT/issues/2)
- [ ] [JOSS review: Missing documentation Feng-Ji-Lab/FedIRT#3](https://github.com/Feng-Ji-Lab/FedIRT/issues/3)
- [ ] [JOSS review: Paper Feng-Ji-Lab/FedIRT#4](https://github.com/Feng-Ji-Lab/FedIRT/issues/4)

Some other comments for when I do start reviewing the functionality:
- Since it seems to be an important part of using the package, it might be worth including in the documentation how to open a port, for those less technically inclined.
- When I use the `runserver()` and `runclient()` functions, the Shiny app appears, but with a display saying `Error: could not find function "shell"` which I assume is unintended behavior. My operating system is Linux (Fedora Kinoite); could this perhaps be a culprit?
