+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://feng-ji-lab.github.io/FedIRT/
+template:
+  bootstrap: 5
+
