+++ DESCRIPTION
@@ -25,3 +25,5 @@ Suggests:
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 LazyData: true
+URL: https://feng-ji-lab.github.io/FedIRT
+BugReports: https://github.com/Feng-Ji-Lab/FedIRT/issues
