### 2025-03-14 Feng-Ji-Lab/FedIRT#1 @BiyingZhou
Thank you for pointing this out. We have updated the `LICENSE` plaintext file to include the full MIT license text, ensuring it is correctly detected by JOSS. Let us know if any further modifications are needed.
