### 2025-03-14 Feng-Ji-Lab/FedIRT#2 @BiyingZhou
Thank you for your question.
The data included in this package is simulated and does not contain any real human participant data. We explicitly stated in the package documentation. The datasets are designed purely for demonstration purposes to showcase the functionality of `FedIRT`.
Specifically, the included data illustrates that federated learning produces results identical to centralized estimation, ensuring that the method works correctly while preserving privacy. No real student responses or sensitive information are included in any part of the package.
Please let us know if further clarification is needed.
