### 2025-03-14 Feng-Ji-Lab/FedIRT#3 @BiyingZhou
Thank you for your detailed feedback. We have addressed each of the documentation requirements as follows:

**Statement of Need**
- We have added a clear statement of need at the top of the `README.md`, explaining what `FedIRT` is used for and why federated learning is important for IRT.
- This ensures that visitors to the repository immediately understand the package’s purpose and significance.

**Installation Instructions**
- The installation instructions have been updated to include both CRAN (stable version) and GitHub (development version) installation methods.

**Example Usage**
- The explanation of how to use the Shiny App has been highlighted in the example usage section, as this is the highlight of the package.
- We have set up a `pkgdown` GitHub Pages site at [https://feng-ji-lab.github.io/FedIRT/](https://feng-ji-lab.github.io/FedIRT/) to host documentation, making it easier for users to browse tutorials and examples.

**Automated Tests**
- The `testthat` folder has been reorganized so that all tests are now automatically detected and executed by test_check("FedIRT"). CI frameworks can properly run all tests now.

**Community Guidelines**
- We have added clear contribution guidelines in `README.md`, detailing:
  - How third parties can contribute to the software.
  - How to report issues via GitHub Issues.
  - How to seek support through documentation, issues, or direct contact.
- This aligns with JOSS’s requirements and makes it easier for new users and contributors to engage with the project.

All these updates should significantly improve the package’s usability and documentation. Thanks again for your constructive feedback!
