### 2025-03-14 Feng-Ji-Lab/FedIRT#4 @BiyingZhou
Thank you for your detailed and thoughtful review. We appreciate your feedback and have made the following revisions accordingly:

**Summary**
- We have restructured the summary to clearly state the motivation for `FedIRT` and why federated learning is needed for IRT.
- We explicitly mention that the package enables researchers to analyze educational testing data while preserving privacy, which is a key challenge in this field.
- The revised version is now more direct and jargon-free to better accommodate non-specialist readers.

**Statement of Need**
- We have expanded our explanation of IRT and provided concrete examples of its applications (e.g., standardized testing, psychological assessments, and large-scale surveys).
- The revised section now makes clear who are the target audience, including education and psychometric researchers, social scientists, and school districts.
- We also highlight that `FedIRT` is the first open-source `R` package that performs federated IRT estimation in a truly distributed manner, addressing current limitations in the field regarding practical implementations of federated learning.

**Quality of Writing & Word Count Reduction**
- We removed the "Method" section and condensed it into a brief statement in the "Statement of Need", explicitly mentioning that we use federated stochastic gradient descent (FedSGD) and federated averaging (FedAvg) for parameter estimation.
- The following sections have been removed as they were more suitable for package documentation rather than the paper:
  - "Availability"
  - "Example of the integrated function"
  - "Example of the personscore function"
  - "Example of the personfit function"
- We have retained the "Example of the Shiny App" as you recommended, ensuring readers understand that `FedIRT` provides an actual working federated solution through Shiny.

These changes have significantly streamlined the paper while maintaining clarity and essential details. The revised version is now more focused and aligns better with the JOSS guidelines.
Thanks again for your valuable comments!
