+++ pkgdown.yml
@@ -1,0 +1,8 @@
+pandoc: 3.1.11
+pkgdown: 2.1.1
+pkgdown_sha: ~
+articles: {}
+last_built: 2025-03-14T18:45Z
+urls:
+  reference: https://feng-ji-lab.github.io/FedIRT/reference
+  article: https://feng-ji-lab.github.io/FedIRT/articles
