### 2025-03-16 #7869 @salbalkus
@BiyingZhou it looks like the tests are still referencing hard paths that only exist on your PC. When I run `test_check("FedIRT")` I get an error saying that `C:\\Users\\zby15\\Downloads\\r\\testdata\\dataset1.csv` does not exist in "test829.R". You'll need to convert these filepaths to be relative so that they can be run on any machine, not just your own  (try something like this [StackOverflow example](https://stackoverflow.com/questions/8898469/is-it-possible-to-use-r-package-data-in-testthat-tests-or-run-examples)).
