### 2025-03-18 2e78c7e
remove local tests
