### 2025-03-18 #7869 @BiyingZhou
The files you mentioned are local test files that were previously added to `.Rbuildignore` and are not part of the R package. However, running `test_check("FedIRT")` still attempted to execute these tests. To resolve this, I have now removed those test files, and the corresponding tests that use package-internal data have been moved to other test files.
