### 2025-03-28 Feng-Ji-Lab/FedIRT#7 @salbalkus
JOSS review: fedirt_file function bugs
Raised as part of JOSS review https://github.com/openjournals/joss-reviews/issues/7869.

When attempting to reproduce functionality of the package, I ran into two small issues with the `fedirt_file` function. First, running the following code gives an error:

```
foo = data.frame(
  X1 = c(1, 1, 1, 0, 0, 0),
  X2 = c(1, 0, 0, 0, 1, 1),
  X3 = c(1, 0, 0, 1, 1, 0),
  S = c("school1","school1","school1","school2","school2","school2")
)

result <- fedirt_file(foo, model_name = "2PL", colname = "S")
```

This is because line 26 in fedirt_file.R assumes the column identifying each site is the first column; instead, this line should select all columns except for this included by name. In addition, it may be of interest to avoid including a default value for the the `colname` parameter in `fedirt_file`; the example in the docs assumes `dataset.csv` has the schools denoted by the default "site" name but this could be confusing if other users try to run it on their own data where this is not the case.

Also, the code below seems to be inconsistent with the documentation:

```
foo = data.frame(
  S = c("school1","school1","school1","school2","school2","school2"),
  X1 = c(1, 1, 1, 0, 0, 0),
  X2 = c(1, 0, 0, 0, 1, 1),
  X3 = c(1, 0, 0, 1, 1, 0)
)

result <- fedirt_file(foo, model_name = "2PL", colname = "S")

summary(result)
            Length Class  Mode
par          6     -none- numeric
value        1     -none- numeric
counts       2     -none- numeric
convergence  1     -none- numeric
message      0     -none- NULL
hessian     36     -none- numeric
loglik       1     -none- numeric
b            3     -none- numeric
a            3     -none- numeric
person       6     -none- list
SE           2     -none- list
```

From the docs I see `summary` should be returning numerical values. A similar result occurs when calling `summary` on the results of `personfit` and `personscore`
