### 2025-03-28 #7869 @salbalkus
Hi @BiyingZhou

Apologies for the delay -- thank you for fixing the test folder, `test_check` seems to work now. I just started evaluating the functionality of the package. In addition to the issue I mentioned above, I have an important question.

To my understanding, the way this package works is that clients run a Shiny app that connects to a server on the researchers' computer. When all clients are ready, the server begins the federated learning process (using `server.R`), distributing model parameters between the server and each client (which respond via the code in `client.R`) until convergence.

Now, please correct me if I am wrong, as I may be mistaken, but in the files `client.R` and `server.R` I don't see any of the code from the rest of the package being used. For example, the documentation seems to imply that `fedirt` or `fedirt_file` perform the actual calculations in the federated learning process, but neither `client.R` nor `server.R` call these functions (in fact, they actually redefine another function named `fedirt` to do something different than the analogous function in the documentation).

This could be perfectly fine, but if this is the case, then it would appear that the unit tests do not cover the part of the package that actually performs the federated learning. In order to complete my reviewer checklist, these parts would need to be covered by the test suite.

(P.S. I would recommend looking into the [shinytest package](https://shiny.posit.co/r/articles/improve/testing-overview/) to be able to cover the Shiny part of the package with unit tests more easily!)
