### 2025-04-08 Feng-Ji-Lab/FedIRT#8 @salbalkus
JOSS Review: Functionality
Raised as part of JOSS review https://github.com/openjournals/joss-reviews/issues/7869.

I am currently unable to replicate the functionality of the package shown in the documentation.

I attempted to test the Shiny app on my local machine by running two instances of R: one on which I called `runserver()` and the other on which I called `runclient()`. On the client side, I connected to the IP displayed by the server side, and then uploaded a 20-unit CSV file very similar to `example_2PL_1`. The only differences were that the column names were Q1, Q2, etc. instead of V1, V2, etc. and the entries were different (but all still 0 or 1). When I clicked "start" on the server side, the application closed saying:

`Warning: Error in rep: invalid 'times' argument`

This error crashed the Shiny app, preventing any computation from taking place. I tested on both Linux and Windows and the error occurred both times. On Linux (Fedora 40) it also said "0 schools connected" and displayed an error `could not find function "shell"` though the server side would visibly update when I connected. On Windows the interface seems to work correctly, saying "1 school connected", though I still received the `invalid 'times' argument` error above. Let me know if you'd like any additional information; I would be happy to oblige.
