### 2025-04-08 Feng-Ji-Lab/FedIRT#8 @salbalkus
Just to add more information, I also saved `example_2PL_1` directly as a CSV and tried to run the Shiny app to try and see if the issue was with my creation of a custom dataset. I saved it in three ways: with the header and row labels, with the header and no row labels, and without the header or row labels. When I uploaded either data with the header, I received the error:

`Warning in ui(req) : NAs introduced by coercion
Warning: Error in if: missing value where TRUE/FALSE needed`

and when I uploaded without the header the server just closed without any results. Clicking "receive results" on the client also closed the client saying `Warning: Error in observe: object 'fedresult' not found`
