### 2025-04-08 #7869 @salbalkus
@BiyingZhou it seems as though the Shiny app contains a bug that prevents me from running it despite my best efforts, as I've described in the issue above. Please let me know if this is able to be addressed; happy to do any more testing on my end or answer any more questions about my setup to assist with bug-fixing.
