+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-04-08T15:56Z
+last_built: 2025-04-09T04:17Z
 urls:
   reference: https://feng-ji-lab.github.io/FedIRT/reference
   article: https://feng-ji-lab.github.io/FedIRT/articles
