### 2025-04-16 Feng-Ji-Lab/FedIRT#7 @BiyingZhou
Thank you for your helpful feedback!
Regarding the `fedirt_file()` function, you are absolutely right—the earlier implementation assumed that the site column had to be numeric. We’ve now updated the code so that the site column can be specified by name and can also be a character vector (e.g., "school1", "school2"). This should resolve the issue you encountered with the column selection and default assumptions. We’ve also taken your suggestion to make the `colname` parameter explicit to avoid confusion for users applying the function to their own data.
As for the `summary()` output, please make sure you're installing and loading the package using the following commands:
``` r
devtools::install_github("Feng-Ji-Lab/FedIRT")
library(FedIRT)
```
Cloning the repo locally may lead to missing compiled objects or outdated code. Installing via `devtools` ensures that all necessary components are properly built and loaded.
Let us know if the issue persists after updating!
