### 2025-04-16 Feng-Ji-Lab/FedIRT#8 @BiyingZhou
Thank you very much for the detailed report!
The previous version of the Shiny app primarily focused on compatibility with Windows systems and did not properly handle IP retrieval or process management in Linux environments. We've since updated the code to support both Linux and Windows by using OS-specific commands for IP address detection and process control. This should resolve the shell-related errors and the "0 schools connected" message on Linux.
As for the error on Windows (Warning: Error in rep: invalid 'times' argument), I have not been able to reproduce this issue during testing. To better understand the situation, could you please provide additional details? Specifically:
The exact sequence of actions you performed (e.g., did you click "Start" before or after all clients connected and uploaded data?)
The structure of the dataset you uploaded (a sample file or the actual CSV would be very helpful)
Whether the app crashed immediately upon clicking "Start", or after some delay?
This information will help us replicate and resolve the issue more efficiently. Thanks again for your help and for testing across different systems!
