### 2025-04-16 #7869 @BiyingZhou
Thank you for the thoughtful question and for continuing to test the package functionality!
You are correct in noting that the `client.R` and `server.R` files used in the Shiny app appear to redefine the `fedirt` function. In fact, these functions are essentially modified versions of the core federated learning logic from the package—copied and adapted to enable network communication and interactive usage within the Shiny framework. While they may not directly call `fedirt()` or `fedirt_file()` as-is, the underlying computational logic is preserved.
Because the actual computation is adapted from the same codebase that is covered by the existing unit tests, the core functionality has already been tested. That said, I understand your concern about ensuring coverage of the Shiny application itself.
Unfortunately, packages like `shinytest` or `shinytest2` are not designed to test apps with remote network communication or multi-session behavior, so they cannot be applied to this use case. However, I’ve added test cases that verify whether the Shiny app components (server and client) can start up and run properly in isolation, which helps ensure basic operability.
If you have suggestions for other testing approaches that might better capture this kind of distributed architecture, I’d be very happy to incorporate them. Thanks again for your detailed feedback!
