### 2025-04-28 Feng-Ji-Lab/FedIRT#8 @salbalkus
Hi @BiyingZhou the latest version that I installed from GitHub seems to fix the Linux issue `could not find function "shell"` when calling `runserver()` but not when calling `runclient()`. I've attached the CSV files I've been using to test.

Here's my sequence of actions on both Windows and Linux for the `Warning: Error in rep: invalid 'times' argument`:

1. Open a session of RStudio
2. Call `library(FedIRT)` and then `runserver()`
3. Open a second session of RStudio
4. Call `library(FedIRT)` and then `runclient()`
5. Copy the IP listed in `Current IP in using` from the server window to the client window and click `Reconnect to Server`
6. Upload "[school1.csv](https://github.com/user-attachments/files/19941026/school1.csv)" to the client by clicking `Browse` on the client window and selecting the file from my desktop.
7. Click `Start` on the server window

The same thing seems to occur when I do the same thing as the above but I open two sessions of the client and upload "[school1.csv](https://github.com/user-attachments/files/19941026/school1.csv)" and "[school2.csv](https://github.com/user-attachments/files/19941027/school2.csv)". I also tried the same thing with a CSV that doesn't have a header, "[example.csv](https://github.com/user-attachments/files/19941028/example.csv)", and this gives me a different error on Linux: `Warning: Error in curl::curl_fetch_memory: Could not resolve host: logL`. I haven't had the chance to re-try with "example.csv" on Windows yet, but I'll let you know once I do.
