### 2025-04-28 Feng-Ji-Lab/FedIRT#7 @salbalkus
Hi @BiyingZhou after installing from GitHub, the `fedirt_file` command now works great, thanks for updating this! On my machine it still looks like `summary(result)` returns the same thing as it did above rather than numerical values though.
