### 2025-04-28 Feng-Ji-Lab/FedIRT#9 @salbalkus
JOSS Review: Test failing
Raised as part of JOSS review https://github.com/openjournals/joss-reviews/issues/7869.

It looks like when I update to the latest version of GitHub and run the tests locally, there seems to be one that is failing. I get the following error:

```
> test_local(path="/home/salbalkus/Research/FedIRT/")
✔ | F W  S  OK | Context
✔ |          6 | 1PL
✔ |          6 | 2PL [1.7s]
✔ |         11 | data
✔ |          4 | graded [55.3s]
✖ | 1        1 | personscore [1.7s]
────────────────────────────────────
Error (test_personscore.R:20:1): (code run outside of `test_that()`)
<subscriptOutOfBoundsError/error/condition>
Error in `inputdata[[colname]]`: subscript out of bounds
Backtrace:
    ▆
 1. └─FedIRT::fedirt_file(as.matrix(example_data_2PL_1), model_name = "2PL") at test_personscore.R:20:1

```

This may be related to recent fixes involving column name/ordering.
