### 2025-04-28 #7869 @salbalkus
Hi @BiyingZhou I appreciate this response. Regarding `shinytest` I think what you have now is good -- ensuring the app runs is all I think is necessary for that part of the code. Generally from my understanding of network programming, code is tested by calling functions locally on mock responses that might appear from the client, so in my opinion there is no need to further test the shiny app components directly.

That said, I do think it is necessary to separate the computational logic from the networking logic. One major purpose of unit tests is to catch issues that might arise when changes or additions are made. Currrently, if someone wants to add to or change the computational logic that has nothing to do with networking, they need to edit code twice -- and if they make a mistake editing the part that is not tested, this will go undetected and break the package.

To solve this, would it be possible to define Shiny app functions like this?

```
fedirt_network <- function(...){
    *networking logic*
    result <- fedirt(...)    # function covered by unit test
    *more networking logic*
}
```

This way computational logic is not duplicated -- it is all in one place, covered by unit tests -- but the complexities of handing network behavior are avoided.
