### 2025-04-30 Feng-Ji-Lab/FedIRT#8 @BiyingZhou
Thanks for the detailed steps and for sharing the CSV files!

The `shell` issue when calling `runserver()` on Linux has been fixed in the latest version, and overall robustness on Linux has been improved.

Your usage of the Shiny app—starting `runserver()` and `runclient()` in separate RStudio sessions and uploading datasets—is completely correct. I followed the same procedure with your datasets ("school1.csv", "school2.csv", and "example.csv") and tested on both Windows and Linux environments.

On both platforms, the application ran successfully and I did not encounter the `Warning: Error in rep: invalid 'times' argument` or the `curl::curl_fetch_memory` error.

If you're still seeing these issues, please double-check:
- That all clients are uploading valid CSVs (including proper headers when required),
- That the latest version of the package is installed on all RStudio sessions,
- And that your network settings allow communication between client and server.

Let me know if the problem persists or if you have more datasets I can help test!
