### 2025-05-05 #7869 @asmitamitra
Hi @BiyingZhou, thanks for making the changes to the test files as suggested @salbalkus. I was able to run the application and run most of the tests. There are a few things I would like to bring to attention -
1. The file fedirt_gpcm_data.R contains several commented code blocks Line 91-168, Line 190-Line253, is that intentional? Can someone were to uncomment and test these files would it work as expected? If yes, can we add some comments on these methods, or if not, would it best to remove these blocks? I noticed the commented methods `g_logL_entry` already exist in the 1_P model.
2. Another point, in the shiny app, server.R on the API /connect, if ip is not found the application should throw 404L as the status code precisely. 400 Status code error indicates that the server was unable to understand the request - could be incorrect syntax or malformed input. 404 is more appropriate here, as it's a not found resource error.
