### 2025-05-06 Feng-Ji-Lab/FedIRT#7 @BiyingZhou
Thanks for the feedback! You're right that sometimes `summary()` may return the default output rather than the expected values from the custom method. This can occasionally happen during package development. However, I just re-tested on both Windows and Linux, and in both environments, the `summary()` function correctly dispatched to the defined method without any issues.

If you're still encountering this problem, could you please share the exact steps you followed to install the package? Sometimes installation issues—especially with development versions—can cause method bindings not to update properly. It might be worth checking if the environment on your side is using the updated code—perhaps a restart of the R session or re-installation from GitHub might help.
