### 2025-05-07 Feng-Ji-Lab/FedIRT#9 @BiyingZhou
Thank you for raising this issue.
I've tested the latest version of the package multiple times locally but have not been able to reproduce the error you're encountering. Could you please confirm that your local repository is cloned from the latest version on GitHub? Before reinstalling, make sure to remove any previously installed versions of the package to avoid conflicts, especially if both a development version and an installed version from CRAN are present in the default library path. This might be the cause of the outdated code being used during testing.

Let me know if the issue persists after a clean installation.
