+++ R/fedirt_gpcm_data.R
@@ -58,18 +58,12 @@ fedirt_gpcm_data = function(inputdata) {
       # 提供所有答案的概率:  4:21
       px_sum = Px_sum(a,b)
       sum_px_sum = matrix(colSums(px_sum), nrow = 1)
-      # if(j==1) {
-      #   ans = broadcast.divide(px_sum, sum_px_sum)
-      #   # print(ans)
-      # }
+
       return(broadcast.divide(px_sum, sum_px_sum))
     })
     log_Lik_j = mem(function(a, b, j) {
-      # 根据答案 data 选对应的概率
-      # 原来： N : 21 = N:10 * 10:21
-      # 现在： N : 21 = 10 * (N:1 select 3:21)
+
       answerP = log(Pjx(a[j], b[[j]]))
-      # 初始化一个全0的矩阵，矩阵尺寸为N行和M[j]列
       result_matrix <- matrix(0, nrow = N[[index]], ncol = M[j] + 1)
       result_matrix[cbind(seq_len(N[[index]]), data[,j] + 1)] = 1
       selected = result_matrix %*% answerP
@@ -87,92 +81,14 @@ fedirt_gpcm_data = function(inputdata) {
     }
     sum(log(matrix(apply(broadcast.multiplication(exp(finalLogL), t(A)), c(1), sum))))
   }
-#
-#   g_logL_gpcm = function(a, b, index, q = 21, lower_bound = -3, upper_bound = 3) {
-#     # init
-#     data = my_data[[index]]
-#     level_diff = (upper_bound - lower_bound) / (q - 1)
-#     X = as.matrix(as.numeric(map(1:q, function(k) {
-#       temp = (lower_bound + (k - 1) * level_diff)
-#       return(temp)
-#     })))
-#     A = as.matrix(as.numeric(map(1:q, function(k) {
-#       temp = (lower_bound + (k - 1) * level_diff)
-#       quadrature = quadl(g, temp - level_diff * 0.5, temp + level_diff * 0.5)
-#       return(quadrature)
-#     })))
-#
-#     Px = mem(function(a, b) {
-#       - rbind(rep(0, length(X)), a * broadcast.subtraction(t(b), t(X)))
-#     })
-#     Px_sum = mem(function(a, b) {
-#       exp(apply(Px(a,b),2,cumsum))
-#     })
-#
-#     Pjx = mem(function(a, b, j) {
-#       # 提供所有答案的概率:  4:21
-#       px_sum = Px_sum(a,b)
-#       sum_px_sum = matrix(colSums(px_sum), nrow = 1)
-#       # if(j==1) {
-#       #   ans = broadcast.divide(px_sum, sum_px_sum)
-#       #   # print(ans)
-#       # }
-#       return(broadcast.divide(px_sum, sum_px_sum))
-#     })
-#     log_Lik_j = mem(function(a, b, j) {
-#       # 根据答案 data 选对应的概率
-#       # 原来： N : 21 = N:10 * 10:21
-#       # 现在： N : 21 = 10 * (N:1 select 3:21)
-#       answerP = log(Pjx(a[j], b[[j]]))
-#       # 初始化一个全0的矩阵，矩阵尺寸为N行和M[j]列
-#       result_matrix <- matrix(0, nrow = N, ncol = M[j] + 1)
-#       result_matrix[cbind(seq_len(N), data[,j] + 1)] = 1
-#       selected = result_matrix %*% answerP
-#       return(selected)
-#     })
-#
-#     Lik_j = mem(function(a, b, j) {
-#       exp(log_Lik_j(a,b,j))
-#     })
-#     LA = mem(function(a, b) {
-#       broadcast.multiplication(Lik(a,b), t(A))
[…]
