+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-04-28T22:29Z
+last_built: 2025-05-07T16:01Z
 urls:
   reference: https://feng-ji-lab.github.io/FedIRT/reference
   article: https://feng-ji-lab.github.io/FedIRT/articles
