### 2025-05-07 Feng-Ji-Lab/FedIRT#9 @salbalkus
Just removed my existing install and tested the latest version, it looks like the test runs correctly now! Thanks for this, feel free to close the issue.
