### 2025-05-07 Feng-Ji-Lab/FedIRT#7 @salbalkus
Hi @BiyingZhou to install the package I just used `remove.packages("FedIRT")` and then called `remotes::install_github("https://github.com/Feng-Ji-Lab/FedIRT")` and `library(FedIRT)` (and then the code in my original issue). It looks like the summary works if I run the testing suite before calling the code in the original issue, but not if I just run the code directly.
