### 2025-05-07 Feng-Ji-Lab/FedIRT#7 @BiyingZhou
Hi @salbalkus ,

Thank you for pointing this out, and I’m very sorry for the confusion. The issue occurred because the `fedirt_file()` function did not assign the appropriate class to its returned object, even though it internally called a function that produced a valid `fedirt` object. As a result, the `summary()` method was not properly dispatched.

I've carefully reviewed and updated the code to ensure that all relevant functions now explicitly assign the `"fedirt"` class to their output where necessary. This should fully resolve the issue, and `summary()` should now work consistently regardless of whether you run the test suite or call the functions directly.

Please let me know if you encounter any further issues!
