### 2025-05-07 #7869 @BiyingZhou
Hi @salbalkus, your suggestion makes a lot of sense, and I really appreciate your thoughtful feedback.
I agree that separating computational logic from networking logic is an important goal for maintainability and testability. We have tried our best to address this issue by moving most of the individual value computations into separate functions without distorting the main function logic. These functions live outside the main Shiny app body and are structured specifically for unit testing.
