### 2025-05-27 Feng-Ji-Lab/FedIRT#8 @salbalkus
Hi @BiyingZhou just tested this again on Linux using the latest version of the package. I no longer receive the `Warning: Error in curl::curl_fetch_memory` error, but now when I run `runclient`  I get an error `object 'sys_name' not found` that displays in the Shiny App.

After that, I'm not sure how long the federated learning process takes, but when I upload either "school1.csv" or "example.csv" and click "Start" on the server, the server doesn't seems to terminate after several minutes. Clicking "receive result" while this process is ongoing throws `Warning: Error in observe: object 'fedresult' not found`. I'll try to look more in detail later today.
