### 2025-05-27 Feng-Ji-Lab/FedIRT#7 @salbalkus
`summary` works now, thanks for fixing this! Feel free to close the issue.
