### 2025-05-27 #7869 @salbalkus
Hi @BiyingZhou,

I just have two remaining concerns:

1. I'm still having trouble evaluating the Shiny app, will continue trying this week. See the corresponding issue in the repo; I ran it again with the latest updates on Linux but not yet Windows; will update.

2. From a testing standpoint, maybe I am mistaken, but I still see most of the main methods of the packages are redefined in https://github.com/Feng-Ji-Lab/FedIRT/blob/main/inst/shiny/server.R and client.R and therefore are not being covered by the tests, which is necessary for me to check the "Automated testing" box. For example, `fedirt` is completely redefined in the Shiny part of the codebase, when Shiny should be calling the method from the `R` folder that is covered by the test. Does that make sense? I know that R cannot test Shiny directly, but almost all of the methods in the Shiny app are just redefinitions of purely mathematical (i.e. non-networking) methods and you already have in the `R` folder. Can you move these methods to the `R` folder so they are covered?
