### 2025-05-28 c39b902
move shiny app functions to utils
