+++ R/server_utils.R
@@ -1,0 +1,424 @@
+fedirt_server = function(J,logL_entry, g_logL_entry) {
+  get_new_ps = function(ps_old) {
+    # "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"
+    optim(par = ps_old, fn = logL_entry, gr = g_logL_entry, method = "BFGS", control = list(fnscale=-1, trace = 0,  maxit = 10000))
+  }
+  ps_init = c(rep(1, J), rep(0, J))
+  ps_next = get_new_ps(ps_init)
+  ps_next$loglik = logL_entry(ps_next$par)
+
+  ps_next$b = ps_next$par[(J+1):(J+J)]
+  ps_next$a = ps_next$par[1:J]
+
+  ps_next
+}
+
+broadcast.fortmat <- function(mat1, mat2) {
+  row1 <- nrow(mat1)
+  col1 <- ncol(mat1)
+  row2 <- nrow(mat2)
+  col2 <- ncol(mat2)
+  if(col1 != 1 && row2 != 1) {
+    stop("illegal operation: not 1")
+  }
+  if(col1 == 1) {
+    mat1_new <- mat1[, rep(1:col1, col2)]
+  } else if(col1 != col2) {
+    stop("illegal operation: col1")
+  } else {
+    mat1_new <- mat1
+  }
+  if(row2 == 1) {
+    mat2_new <- mat2[rep(1:row2, each=row1), ]
+  } else if(row2 != row1) {
+    stop("illegal operation: row2")
+  } else {
+    mat2_new <- mat2
+  }
+
+  list(mat1_new, mat2_new)
+}
+
+broadcast.multiplication <- function(mat1, mat2) {
+  format_result = broadcast.fortmat(mat1, mat2)
+  mat1_new = format_result[[1]]
+  mat2_new = format_result[[2]]
+  return(mat1_new * mat2_new)
+}
+broadcast.divide <- function(mat1, mat2) {
+  format_result = broadcast.fortmat(mat1, mat2)
+  mat1_new = format_result[[1]]
+  mat2_new = format_result[[2]]
+  return(mat1_new / mat2_new)
+}
+broadcast.subtraction <- function(mat1, mat2) {
+  format_result = broadcast.fortmat(mat1, mat2)
+  mat1_new = format_result[[1]]
+  mat2_new = format_result[[2]]
+  return(mat1_new - mat2_new)
+}
+broadcast.exponentiation <- function(mat1, mat2) {
+  format_result = broadcast.fortmat(mat1, mat2)
+  mat1_new = format_result[[1]]
+  mat2_new = format_result[[2]]
+  return(mat1_new ^ mat2_new)
+}
+mem <- function(f) {
+  memo <- new.env(parent = emptyenv())
+  function(...) {
+    key <- paste(list(...), collapse = " ,")
+    if(!exists(as.character(key), envir = memo)) {
+      memo[[as.character(key)]] <- f(...)
+    }
+    memo[[as.character(key)]]
+  }
+}
+g = function(x) {
+  return (exp(-0.5 * x * x) / sqrt(2 * pi))
+}
+
+logL = function(a, b, data, q = 21, lower_bound = -3, upper_bound = 3) {
+  # init
+  N = nrow(data)
+  J = dim(data)[2]
+  level_diff = (upper_bound - lower_bound) / (q - 1)
+  X = as.matrix(as.numeric(map(1:q, function(k) {
+    index = (lower_bound + (k - 1) * level_diff)
+    return(index)
+  })))
[…]

[…]
