### 2025-05-28 #7869 @BiyingZhou
Hi @salbalkus,

Thank you for your thoughtful feedback.

I’ve now updated the Shiny app as per your suggestions. Specifically:

- I’ve removed all duplicate definitions of functions that are already covered in the `R/` folder from both `server.R` and `client.R`.

- For functions in the Shiny app that had name conflicts with functions in the `R/` folder but required modified behavior, I’ve renamed them to avoid ambiguity and ensure clarity in the codebase.

With these changes, the Shiny app now calls functions defined in the `R/` folder directly, ensuring these core functionalities are exercised and covered by the existing automated tests. This should address the concern about test coverage for mathematical methods used in the Shiny app.

I have re-tested the entire package following these updates on both Windows and Linux.

Please feel free to let me know if anything still looks off. I really appreciate your careful review and support throughout this process.
