### 2025-06-06 #7869 @salbalkus
Thanks for this update @BiyingZhou. When I install the package's latest version from GitHub and load it using
```
devtools::install_github("Feng-Ji-Lab/FedIRT")
library(FedIRT)
```
and I run the package now (taking the same steps I described in [this issue](https://github.com/Feng-Ji-Lab/FedIRT/issues/8)), pressing "Start" on the server application throws the error `"Warning: Error in fedirt_server: could not find function "fedirt_server"` on both Windows and Linux. It seems like `fedirt_server` is not exported in NAMESPACE so it's probably worth checking that again and making sure all of the functions you've moved into the package are properly exported. Or try calling the function using `FedIRT:::fedirt_server` (and the same for other functions from your package) to see if that is sufficient to resolve the issue.
