### 2025-06-07 76aead8
modify client and server utils
