### 2025-06-07 #7869 @BiyingZhou
Hi @salbalkus,

Thank you for pointing this out. I’ve investigated the issue and realized that `fedirt_server` and related functions were not exported in the `NAMESPACE` file as they were intended for internal use in the Shiny app. Following your suggestion, I’ve revised the package so that these internal functions can now be accessed via `FedIRT:::` and continue not to be exported. This resolves the issue when launching the Shiny app.

The fix has been committed, and the latest version of the package on GitHub now works as expected. Let me know if you run into any further problems. Thank you again for your thorough review!
