### 2025-06-07 Feng-Ji-Lab/FedIRT#8 @salbalkus
After the latest update, I was able to successfully run the app on Linux using `example.csv`. Great work! I have one further comment: Referring to my previous comment, I ran this package twice: once using `example.csv` and once using `school1.csv`  and `school2.csv`. The package worked when I used `example.csv` but when I run it using `school1.csv` and `school2.csv` the server crashed with the error `Warning: Error in rep: invalid 'times' argument` (the same happens even if I just use `school1.csv`). I'm not sure why this happens, but could it be due to how `school1.csv` is structured differently? Do you mind taking a look and seeing if there is an issue with how this CSV is formatted? Thanks!
