### 2025-06-07 Feng-Ji-Lab/FedIRT#8 @BiyingZhou
Thanks for the update and for testing with different files!

I’ve looked into the issue and identified that the crash was caused by incorrect header handling when using files like `school1.csv`. This has now been fixed — the app can correctly detect and process headers, so you should no longer encounter the `invalid 'times'` argument error.

Please try again with your files and let us know if everything runs smoothly now. Thanks again for your thorough testing!
