### 2025-06-08 Feng-Ji-Lab/FedIRT#8 @salbalkus
Thanks for updating this -- I can now run the app successfully 👍 Once I have time later in the week I'll test everything for correctness and make sure the examples in the documentation reproduce, and if everything goes smoothly I'll finish off my checklist.
