### 2025-06-16 #7869 @BiyingZhou
Hi @salbalkus,
Thanks so much for the thoughtful feedback and for verifying the 2PL results — I really appreciate it!

Regarding your question, the Shiny App supports the graded model. You just need to upload a dataset formatted for the graded model when using the app. I’ve also updated the README to clarify this, so users are aware of how to trigger graded model estimation through data input.

Let me know if anything is unclear or if you run into any issues!
