### 2025-06-16 #7869 @salbalkus
> Hi [@salbalkus](https://github.com/salbalkus), Thanks so much for the thoughtful feedback and for verifying the 2PL results — I really appreciate it!
>
> Regarding your question, the Shiny App supports the graded model. You just need to upload a dataset formatted for the graded model when using the app. I’ve also updated the README to clarify this, so users are aware of how to trigger graded model estimation through data input.
>
> Let me know if anything is unclear or if you run into any issues!

Ah I see, that makes sense. Nice that it's automatic!

I just tested the app on `example_data_graded.csv` and I did run into a minor issue -- the model would converge to a specific value of the loss, but the fitting algorithm would never actually stop, it would just keep iterating and oscillating around the loss function (at least, after running it for a few minutes). Could there be a way to stop the model fitting process manually, or at least stop it after something like 500 iterations?
