+++ DESCRIPTION
@@ -4,10 +4,10 @@ Title: Federated Item Response Theory Models
 Version: 1.2.0
 Authors@R: c(
     person("Biying", "Zhou", email = "zby.zhou@mail.utoronto.ca", role = c("cre","aut")),
-    person("Feng", "Ji", email = "f.ji@utoronto.ca", role = "aut"),
-    person("Nanyu", "Luo", email = "n.luo@mail.utoronto.ca", role = "aut"),
     person("Yuhao", "Yuan", email = "yuhao.yuan@mail.utoronto.ca", role = "aut"),
-    person("Jia", "Qi", email = "jiajia.qi@mail.utoronto.ca", role = "aut"))
+    person("Nanyu", "Luo", email = "n.luo@mail.utoronto.ca", role = "aut"),
+    person("Jia", "Qi", email = "jiajia.qi@mail.utoronto.ca", role = "aut"),
+    person("Feng", "Ji", email = "f.ji@utoronto.ca", role = "aut"))
 Description: Integrate Item Response Theory (IRT) and Federated Learning to estimate traditional IRT models, including the 2-Parameter Logistic (2PL) and the Graded Response Models, with enhanced privacy. It allows for the estimation in a distributed manner without compromising accuracy. A user-friendly 'shiny' application is included. 
 License: MIT + file LICENSE
 Depends:
