### 2025-08-01 #7869 @BiyingZhou
Hi @samhforbes,
Here is the DOI for our package v1.3.0: https://doi.org/10.5281/zenodo.16691212
Please let us know if there are any further changes to be made!
