+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: FedIRT
 Title: Federated Item Response Theory Models
-Version: 1.2.0
+Version: 1.3.0
 Authors@R: c(
     person("Biying", "Zhou", email = "zby.zhou@mail.utoronto.ca", role = c("cre","aut")),
     person("Yuhao", "Yuan", email = "yuhao.yuan@mail.utoronto.ca", role = "aut"),
