+++ paper/paper.bib
@@ -36,10 +36,14 @@
     organization={University of Toronto},
     address={Toronto, Canada}
 }
-@unpublished{FederatedIRT2024_1,
-    title = {Federated Item Response Models},
-    author = {Biying Zhou and Feng Ji},
-    year = {In submission}
+@misc{zhou2025federateditemresponsetheory,
+      title={Federated Item Response Theory Models}, 
+      author={Biying Zhou and Nanyu Luo and Feng Ji},
+      year={2025},
+      eprint={2506.21744},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      url={https://arxiv.org/abs/2506.21744}, 
 }
 
 @article{chalmers2012mirt,
