### 2025-08-06 #7869 @BiyingZhou
Hi @samhforbes,
We have updated the JOSS paper as shown in the repository now. Are we ready to proceed to the next steps, or are there still some tasks to be done?
Thank you for your time to review our paper!
