+++ LICENSE.md
@@ -1,6 +1,6 @@
 # MIT License
 
-Copyright (c) 2024 Biying Zhou and Feng Ji
+Copyright (c) 2024 Biying Zhou, Yuhao Yuan, Nanyu Luo, Jia Qi and Feng Ji
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
