### 2025-08-16 #7869 @BiyingZhou
> Hi [@BiyingZhou](https://github.com/BiyingZhou) I can see the licence in the archive should match the one for the repository, and the title and author list should also match here.

Hi @samhforbes,
We have updated the license for the repository and ensured that the author list matches that in the paper.
