### 2025-08-16 #7869 @BiyingZhou
> [@BiyingZhou](https://github.com/BiyingZhou) both in Chalmers and in Rizopoulos the software name that should be capitalised (such as "R" rather than "r") needs to be formatted correctly

Thank you for pointing this out. I double-checked the `.bib` file, and the capitalization is correct there (e.g., "R" instead of "r"). However, the generated PDF still displays it incorrectly. I’m using the compilation command provided by JOSS, and it seems that the capitalization formatting is being overridden in the output.

Do you have suggestions on how I could fix this issue within the JOSS workflow?
