### 2025-08-20 d714d5d
modify citations
