+++ paper/paper.bib
@@ -47,7 +47,7 @@
 }
 
 @article{chalmers2012mirt,
-  title={mirt: A multidimensional item response theory package for the R environment},
+  title={{mirt}: A multidimensional item response theory package for the {R} environment},
   author={Chalmers, R Philip},
   journal={Journal of Statistical Software},
   volume={48},
@@ -56,7 +56,7 @@
   doi = {10.18637/jss.v048.i06}
 }
 @article{rizopoulos2007ltm,
-  title={ltm: An R package for latent variable modeling and item response analysis},
+  title={{ltm}: An {R} package for latent variable modeling and item response analysis},
   author={Rizopoulos, Dimitris},
   journal={Journal of Statistical Software},
   volume={17},
@@ -78,7 +78,7 @@
   doi = {10.48550/arXiv.1610.05492}
 }
 @article{fox2007multilevel,
-  title={Multilevel IRT modeling in practice with the package mlirt},
+  title={Multilevel IRT modeling in practice with the package {mlirt}},
   author={Fox, Jean-Paul},
   journal={Journal of Statistical Software},
   volume={20},
@@ -87,7 +87,7 @@
   doi={10.18637/jss.v020.i05}
 }
 @article{bock1981marginal,
-  title={Marginal maximum likelihood estimation of item parameters: Application of an EM algorithm},
+  title={Marginal maximum likelihood estimation of item parameters: Application of an {EM} algorithm},
   author={Bock, R Darrell and Aitkin, Murray},
   journal={Psychometrika},
   volume={46},
