### 2025-08-28 #7869 @BiyingZhou
Hi @samhforbes Thank you for pointing out the mismatched name and license.
I updated the archive, including the title and license. The new DOI is [10.5281/zenodo.16691211](https://doi.org/10.5281/zenodo.16691211) now.
Let me know if further changes should be made!
