### 2025-08-29 #7869 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
